refactor : 테스트를 위한 타입 분리
