+++ TodoMate.xcodeproj/xcuserdata/hs.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -23,7 +23,7 @@
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-			<integer>3</integer>
+			<integer>2</integer>
 		</dict>
 		<key>TodoMate.xcscheme_^#shared#^_</key>
 		<dict>
@@ -40,7 +40,17 @@
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
+		<key>4526B0D52D48F56500DC90F5</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
 		<key>4533EB892C9D5FB40058761F</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>4560E38D2D538DB5008B09AA</key>
 		<dict>
 			<key>primary</key>
 			<true/>
