+++ TodoMate/Service/ChatService.swift
@@ -7,13 +7,6 @@
 
 import Observation
 
-protocol ChatServiceType {
-    func create(_ chat: Chat)
-    func fetch() async -> [Chat]
-    func update(_ chat: Chat)
-    func remove(_ chat: Chat)
-}
-
 final class ChatService: ChatServiceType {
     private let chatRepository: ChatRepositoryType
     
@@ -66,20 +59,3 @@ extension ChatService {
     }
 }
 
-class StubChatService: ChatServiceType {
-    func fetch() async -> [Chat] {
-        Chat.stub
-    }
-    
-    func remove(_ chat: Chat) {
-        
-    }
-    
-    func update(_ chat: Chat) {
-        
-    }
-    
-    func create(_ chat: Chat) {
-        
-    }
-}
